Jumping added, problems with walking state change,
Jumping works as indented, although we do have to go a bit further into making sure that the conditions are true as right now, it is passed as always true for testing. Walking state has a bug that if you hold a either A or D and you go over an edge, it'll keep you floating.